refactor(views): extract MdiArea from TabsShell
No behaviour change — lifts the per-tab terminal layout into a new
MdiArea component so subsequent tasks can swap layout strategies by
viewMode without touching the tab strip.
